requirements.txt
# Deep Learning & Machine Learning
tensorflow>=2.10.0
keras>=2.10.0
scikit-learn>=1.2.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.4.0

# Data Handling
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Web Interface
streamlit>=1.28.0

# Utilities
tqdm>=4.64.0
